Move LicenseMixin and accompanying code to openedx/core/lib.
