+++ common/lib/xmodule/xmodule/video_module/video_module.py
@@ -46,7 +46,7 @@ from .video_xfields import VideoFields
 from .video_handlers import VideoStudentViewHandlers, VideoStudioViewHandlers
 
 from xmodule.video_module import manage_video_subtitles_save
-from xmodule.mixin import LicenseMixin
+from openedx.core.lib.license import LicenseMixin
 
 # The following import/except block for edxval is temporary measure until
 # edxval is a proper XBlock Runtime Service.
