+++ openedx/core/lib/license/__init__.py
@@ -1,0 +1,2 @@
+from .wrapper import wrap_with_license
+from .mixin import LicenseMixin
