+++ openedx/core/lib/license/templates/license.html
