+++ openedx/core/lib/license/templates/license_wrapper.html
